Stop training when neural network return a NaN

+++ src/neural_network/layer/neuron/activation_function/Tanh.hpp
@@ -17,12 +17,12 @@ namespace snn::internal
 
         float function(const float x) const override
         {
-            return std::tanh(x);
+            return tanhf(x);
         }
 
         float derivative(const float x) const override
         {
-            return 1 - powf(std::tanh(x), 2);
+            return 1 - powf(tanhf(x), 2);
         }
     };
 }
